+++ R/fit_percentiles.R
@@ -2,11 +2,11 @@
 #'
 #' @description
 #'
-#' This function estimates the percentiles of weighted time series observations. The output contains the percentiles
-#' from the fitted distribution.
+#' This function estimates the percentiles of weighted time series cases or incidences.
+#' The output contains the percentiles from the fitted distribution.
 #'
-#' @param weighted_observations A tibble containing two columns of length n; `observation`, which contains the data
-#' points, and `weight`, which is the importance assigned to the observation. Higher weights indicate that an
+#' @param weighted_observations A tibble containing two columns of length n; `observation`, which contains either cases
+#' or incidences, and `weight`, which is the importance assigned to the observation. Higher weights indicate that an
 #' observation has more influence on the model outcome, while lower weights reduce its impact.
 #' @param conf_levels A numeric vector specifying the confidence levels for parameter estimates. The values have
 #' to be unique and in ascending order, that is the lowest level is first and highest level is last.
@@ -78,14 +78,34 @@ fit_percentiles <- function(
   checkmate::assert_tibble(weighted_observations, add = coll)
   checkmate::assert_numeric(conf_levels, lower = 0, upper = 1,
                             unique = TRUE, sorted = TRUE, add = coll)
-  checkmate::assert_names(colnames(weighted_observations),
-                          identical.to = c("observation", "weight"), add = coll)
+  checkmate::assert_true(
+    length(weighted_observations) >= 2,
+    add = coll
+  )
+  checkmate::assert_names(
+    colnames(weighted_observations),
+    must.include = "weight",
+    subset.of = c("weight", "cases", "incidence", "observation"),
+    add = coll
+  )
   checkmate::assert_numeric(lower_optim, add = coll)
   checkmate::assert_numeric(upper_optim, add = coll)
   checkmate::reportAssertions(coll)
   # Match the arguments.
   family <- rlang::arg_match(family)
   optim_method <- rlang::arg_match(optim_method)
+
+  # Rename cases or incidence to observation if they are present and observation is not
+  if ("incidence" %in% names(weighted_observations) && !"observation" %in% names(weighted_observations)) {
+    weighted_observations <- weighted_observations |>
+      dplyr::rename(observation = "incidence")
+  }
+  if ("cases" %in% names(weighted_observations) &&
+      !any(c("observation", "incidence") %in% names(weighted_observations))
+  ) {
+    weighted_observations <- weighted_observations |>
+      dplyr::rename(observation = "cases")
+  }
 
   # If there is only one unique observation we cannot optimise -> return NA
   if (length(unique(weighted_observations$observation)) == 1) {