+++ R/seasonal_onset.R
@@ -2,12 +2,13 @@
 #'
 #' @description
 #'
-#' This function performs automated and early detection of seasonal epidemic onsets on a time series dataset.
-#' It estimates growth rates for consecutive time intervals and calculates the sum of cases (sum_of_cases).
-#' If the time series data includes a population column it will be used as offset to adjust the growth rate in the glm.
+#' This function performs automated and early detection of seasonal epidemic onsets on a `tsd` object.
+#' It estimates growth rates and calculates the average sum of cases for consecutive time intervals (`k`).
+#' If the time series data includes `population` it will be used as offset to adjust the growth rate in the glm,
+#' additionally the output will include incidence, population and average sum of incidence.
 #'
 #' @param tsd `r rd_tsd`
-#' @param k An integer specifying the window size for modeling growth rates for the onset.
+#' @param k An integer specifying the window size for modeling growth rates and average sum of cases.
 #' @param level The confidence level for onset parameter estimates, a numeric value between 0 and 1.
 #' @param disease_threshold `r rd_disease_threshold(usage = "onset")`
 #' @param family `r rd_family()`
@@ -23,35 +24,26 @@
 #' @examples
 #' # Create a tibble object from sample data
 #' tsd_data <- tsd(
-#'   observation = c(100, 120, 150, 180, 220, 270),
-#'   time = as.Date(c(
-#'     "2023-01-01",
-#'     "2023-01-02",
-#'     "2023-01-03",
-#'     "2023-01-04",
-#'     "2023-01-05",
-#'     "2023-01-06"
-#'   )),
-#'   time_interval = "day"
+#'   cases = c(100, 120, 150, 180, 220, 270),
+#'   time = seq(from = as.Date("2023-01-01"), by = "1 week", length.out = 6)
 #' )
 #'
-#' # Estimate seasonal onset with a 3-day window and a Poisson family model
+#' # Estimate seasonal onset with a 3-day window
 #' seasonal_onset(
 #'   tsd = tsd_data,
 #'   k = 3,
-#'   level = 0.95,
-#'   disease_threshold = 20,
-#'   family = "poisson",
+#'   level = 0.975,
+#'   disease_threshold = 5,
 #'   na_fraction_allowed = 0.4,
-#'   season_start = NULL,
-#'   season_end = NULL,
-#'   only_current_season = NULL
+#'   season_start = 21,
+#'   season_end = 20,
+#'   only_current_season = FALSE
 #' )
 seasonal_onset <- function(                                     # nolint: cyclocomp_linter.
     tsd,
     k = 5,
     level = 0.95,
-    disease_threshold = NA_integer_,
+    disease_threshold = NA_real_,
     family = c(
       "quasipoisson",
       "poisson"
@@ -66,15 +58,15 @@ seasonal_onset <- function(                                     # nolint: cycloc
   checkmate::assert_class(tsd, "tsd", add = coll)
   checkmate::assert_names(
     colnames(tsd),
-    must.include = c("time", "observation"),
-    subset.of = c("time", "observation", "population"),
+    must.include = c("time", "cases"),
+    subset.of = c("time", "cases", "incidence", "population"),
     add = coll
   )
   checkmate::assert_numeric(level, lower = 0, upper = 1, add = coll)
   checkmate::assert_numeric(na_fraction_allowed, lower = 0, upper = 1,
                             add = coll)
   checkmate::assert_integerish(k, add = coll)
-  checkmate::assert_integerish(disease_threshold, add = coll)
+  checkmate::assert_numeric(disease_threshold, add = coll)
   checkmate::assert_integerish(season_start, lower = 1, upper = 53,
                                null.ok = TRUE, add = coll)
   checkmate::assert_integerish(season_end, lower = 1, upper = 53,
@@ -97,6 +89,16 @@ seasonal_onset <- function(                                     # nolint: cycloc
   } else {
     tsd <- tsd |> dplyr::mutate(season = "not_defined")
   }
+
+  # Use incidence if in tsd else use cases
+  use_incidence <- FALSE
+  if ("population" %in% names(tsd) && "incidence" %in% names(tsd)) {
+    use_incidence <- TRUE
+  }
+
+  # Define observation as cases in `tsd`.
+  tsd <- tsd |>
+    dplyr::mutate(observation = .data$cases)
 
   # Extract only current season if assigned
   if (!is.null(season_start) && only_current_season == TRUE) {
@@ -142,7 +144,7 @@ seasonal_onset <- function(                                     # nolint: cycloc
     obs_iter <- tsd[(i - k + 1):i, ]
 
     # Evaluate NA and zero values in windows
-    if (sum(is.na(obs_iter) | obs_iter == 0) > k * na_fraction_allowed) {
+    if (sum(is.na(obs_iter$observation) | obs_iter$observation == 0) > k * na_fraction_allowed) {
       skipped_window[i] <- TRUE
       # Set fields to NA since the window is skipped
       growth_rates <- list(estimate = c(NA, NA, NA),
@@ -150,7 +152,7 @@ seasonal_onset <- function(                                     # nolint: cycloc
     } else {
       # Estimate growth rates
       growth_rates <- fit_growth_rate(
-        observation = obs_iter$observation,
+        cases = obs_iter$observation,
         population = if ("population" %in% names(tsd)) obs_iter$population else NULL,
         level = level,
         family = family
@@ -160,29 +162,34 @@ seasonal_onset <- function(                                     # nolint: cycloc
     # See if the growth rate is significantly higher than zero
     growth_warning <- growth_rates$estimate[2] > 0
 
-    # Calculate Sum of Cases (sum_of_cases)
-    sum_of_cases <- base::sum(obs_iter$observation, na.rm = TRUE)
-
-    # Evaluate if sum_of_cases exceeds disease_threshold
-    sum_of_cases_warning <- sum_of_cases > (disease_threshold * k)
+    if (use_incidence) {
+      # Calculate average incidence in window (k)
+      average_observations_window <- base::sum(obs_iter$incidence, na.rm = TRUE) / k
+    } else {
+      # Calculate average cases in window (k)
+      average_observations_window <- base::sum(obs_iter$cases, na.rm = TRUE) / k
+    }
+    # Evaluate if average_incidence_window exceeds disease_threshold
+    average_observations_warning <- average_observations_window > disease_threshold
 
     # Give an seasonal_onset_alarm if both criteria are met
-    seasonal_onset_alarm <- growth_warning & sum_of_cases_warning
+    seasonal_onset_alarm <- growth_warning & average_observations_warning
 
     # Collect the results
     res <- dplyr::bind_rows(
       res,
       tibble::tibble(
         reference_time = tsd$time[i],
-        observation = tsd$observation[i],
-        population = if ("population" %in% names(tsd)) tsd$population[i] else NULL,
+        cases = tsd$cases[i],
         season = tsd$season[i],
+        population = if ("population" %in% names(tsd)) tsd$population[i] else NA,
+        incidence = if ("incidence" %in% names(tsd)) tsd$incidence[i] else NA,
         growth_rate = growth_rates$estimate[1],
         lower_growth_rate = growth_rates$estimate[2],
         upper_growth_rate = growth_rates$estimate[3],
         growth_warning = growth_warning,
-        sum_of_cases = sum_of_cases,
-        sum_of_cases_warning = sum_of_cases_warning,
+        average_observations_window = average_observations_window,
+        average_observations_warning = average_observations_warning,
         seasonal_onset_alarm = seasonal_onset_alarm,
         skipped_window = skipped_window[i],
         converged = growth_rates$fit$converged
@@ -208,11 +215,13 @@ seasonal_onset <- function(                                     # nolint: cycloc
     k = k,
     level = level,
     disease_threshold = disease_threshold,
-    family = family
+    family = family,
+    incidence_denominator = attr(tsd, "incidence_denominator")
   )
 
   # Keep attributes from the `tsd` class
   attr(ans, "time_interval") <- attr(tsd, "time_interval")
+  attr(ans, "incidence_denominator") <- attr(tsd, "incidence_denominator")
 
   return(ans)
 }